Create separate files for intersection and concatenation
