+++ include/mata/nfa.hh
@@ -723,139 +723,6 @@ inline Nfa uni(const Nfa &lhs, const Nfa &rhs)
 } // uni }}}
 
 /**
- * Class handling intersection of automata.
- *
- * Implements a normal intersection and an intersection preserving epsilon transitions.
- */
-class Intersection
-{
-public:
-    /**
-     * Compute classic intersection of NFAs @p lhs and @p rhs.
-     * @param lhs First NFA to compute intersection for.
-     * @param rhs Second NFA to compute intersecion for.
-     */
-    Intersection(const Nfa& lhs, const Nfa& rhs) : lhs(lhs), rhs(rhs)
-    {
-        compute();
-    }
-
-    /**
-     * Compute epsilon transitions preserving intersection of NFAs @p lhs and @p rhs.
-     * @param lhs First NFA to compute intersection for.
-     * @param rhs Second NFA to compute intersection for.
-     * @param epsilon Symbol to handle as an epsilon symbol.
-     */
-    Intersection(const Nfa& lhs, const Nfa& rhs, const Symbol epsilon) : lhs(lhs), rhs(rhs), epsilon(epsilon)
-    {
-        compute_preserving_epsilon_transitions();
-    }
-
-    /**
-     * Get the final product NFA of the intersection.
-     * @return Product NFA of the intersection.
-     */
-    const Nfa& get_product() { return product; }
-
-    /**
-     * Get product map for the generated product NFA.
-     * @return Product map mapping original state pairs to new product states.
-     */
-    const ProductMap& get_product_map() { return product_map; }
-private:
-    Nfa product{}; ///< Product of the intersection.
-    /// Product map for the generated intersection mapping original state pairs to new product states.
-    ProductMap product_map{};
-    const Nfa& lhs{}; ///< First NFA to compute intersection for.
-    const Nfa& rhs{}; ///< Second NFA to compute intersection for.
-    const Symbol epsilon{}; ///< Symbol to handle as an epsilon symbol.
-
-    StatePair pair_to_process{}; ///< State pair of original states currently being processed.
-    std::set<StatePair> pairs_to_process{}; ///< Set of state pairs of original states to process.
-
-    /**
-     * Compute classic intersection.
-     */
-    void compute();
-
-    /**
-     * Compute intersection preserving epsilon transitions.
-     */
-    void compute_preserving_epsilon_transitions();
-
-    /**
-     * Initialize pairs to process with initial state pairs.
-     */
-    void initialize_pairs_to_process();
-
-    /**
-     * Add transition to the product.
-     * @param[in] intersection_transition State transitions to add to the product.
-     */
-    void add_product_transition(const TransSymbolStates& intersection_transition);
-
-    /**
-     * Update product with initial state pairs.
-     * @param[in] lhs_initial_state Initial state of NFA @c lhs.
-     * @param[in] rhs_initial_state Initial state of NFA @c rhs.
-     */
-    void handle_initial_state_pairs(State lhs_initial_state, State rhs_initial_state);
-
-    /**
-     * Compute product for state transitions with same symbols.
-     * @param[in] lhs_state_transitions State transitions of NFA @c lhs to compute product for.
-     * @param[in] rhs_state_transitions State transitions of NFA @c rhs to compute product for.
-     */
-    void compute_for_same_symbols(const TransSymbolStates& lhs_state_transitions,
-                                  const TransSymbolStates& rhs_state_transitions);
-
-    /**
-     * Compute product for state transitions with @c lhs state epsilon transition.
-     * @param[in] lhs_state_transitions State transitions of NFA @c lhs to compute product for.
-     */
-    void compute_for_lhs_state_epsilon_transition(const TransSymbolStates& lhs_state_transitions);
-
-    /**
-     * Compute product for state transitions with @c rhs state epsilon transition.
-     * @param[in] rhs_state_transitions State transitions of NFA @c rhs to compute product for.
-     */
-    void compute_for_rhs_epsilon_transition(const TransSymbolStates& rhs_state_transitions);
-
-    // TODO: Keep or finally remove?
-    /* If q is the state of this automaton and p of other, then thisAndOtherStateToIntersectState[q][p] = state in intersect
-     * representing (q,p). The hash function computes for each pair (q,p) unique number (q + p*(num of states in this automaton))
-     * whose std hash is returned.
-     *
-     * see https://stackoverflow.com/questions/15719084/how-to-use-lambda-function-as-hash-function-in-unordered-map
-     */
-    State hashStatePair(const StatePair &sp)
-    {
-        return std::hash<unsigned long>()(sp.first + sp.second*lhs.transitionrelation.size());
-    }
-
-    /**
-     * Compute product for current state pair preserving epsilon transitions.
-     * @param[in] lhs_state_transitions State transitions of NFA @c lhs to compute product for.
-     * @param[in] rhs_state_transitions State transitions of NFA @c rhs to compute product for.
-     */
-    void compute_for_state_pair_preserving_eps_trans(const TransSymbolStates& lhs_state_transitions,
-                                                     const TransSymbolStates& rhs_state_transitions);
-
-    /**
-     * Compute classic product for current state pair.
-     */
-    void compute_for_state_pair();
-
-    /**
-     * Create product state and its transitions.
-     * @param[in] lhs_state_to Target state in NFA @c lhs.
-     * @param[in] rhs_state_to Target state in NFA @c rhs.
-     * @param[out] intersect_transitions Transitions of the product state.
-     */
-    void create_product_state_and_trans(State lhs_state_to, State rhs_state_to, TransSymbolStates& intersect_transitions);
-}; // Intersection
-
-/**
  * @brief Compute intersection of two NFAs preserving epsilon transitions.
  *
  * Create product of two NFAs, where both automata can contain ε-transitions. The product preserves the ε-transitions
@@ -890,101 +757,7 @@ void intersection(Nfa* res, const Nfa& lhs, const Nfa& rhs, ProductMap* prod_map
  * @param[in] rhs Second NFA to compute intersection for.
  * @return NFA as a product of NFAs @p lhs and @p rhs with ε-transitions preserved.
  */
-inline Nfa intersection(const Nfa &lhs, const Nfa &rhs) { return Intersection{ lhs, rhs }.get_product(); }
-
-/**
- * Class executing a concatenation of two NFAs.
- */
-class Concatenate
-{
-public:
-    /**
-     * Initialize and compute concatenation of two NFAs.
-     *
-     * Concatenation will proceed in the order of the passed automata: Result is 'lhs . rhs'.
-     * @param[in] lhs First NFA to concatenate.
-     * @param[in] rhs Second NFA to concatenate.
-     */
-    Concatenate(const Nfa& lhs, const Nfa& rhs)
-        : lhs(lhs), rhs(rhs), lhs_states_num(lhs.get_num_of_states()), rhs_states_num(rhs.get_num_of_states())
-    {
-        result.increase_size(lhs_states_num - lhs.finalstates.size() + rhs_states_num);
-
-        compute_concatenation();
-    }
-
-    /**
-     * Get result of concatenation.
-     * @return Concatenated automaton.
-     */
-    Nfa& get_result() { return result; }
-
-    /**
-     * Get @c lhs to @c result states map.
-     * @return @c lhs to @c result states map.
-     */
-    StateMap<State>& get_lhs_result_states_map() { return lhs_result_states_map; }
-
-    /**
-     * Get @c rhs to @c result states map.
-     * @return @c rhs to @c result states map.
-     */
-    StateMap<State>& get_rhs_result_states_map() { return rhs_result_states_map; }
-
-private:
-    const Nfa& lhs{}; ///< First automaton to concatenate.
-    const Nfa& rhs{}; ///< Second automaton to concatenate.
-    const unsigned long lhs_states_num{}; ///< Number of states in @c lhs.
-    const unsigned long rhs_states_num{}; ///< Number of states in @c rhs.
-    Nfa result{}; ///< Concatenated automaton.
-    StateMap<State> lhs_result_states_map{}; ///< Map mapping @c lhs states to @c result states.
-    StateMap<State> rhs_result_states_map{}; ///< Map mapping @c rhs states to @c result states.
-
-    /**
-     * Compute concatenation of given automata.
-     */
-    void compute_concatenation();
-
-    /**'
-     * Map @c lhs and @c rhs states to @c result states.
-     */
-    void map_states_to_result_states();
-
-    /**
-     * Make @c result initial states.
-     */
-    void make_initial_states();
-
-    /**
-     * Make @c result final states.
-     */
-    void make_final_states();
-
-    /**
-     * Add @c rhs transitions to the @c result.
-     */
-    void add_rhs_transitions();
-
-    /**
-     * Add @c lhs transitions from final states to the @c result.
-     */
-    void add_lhs_final_states_transitions();
-
-    /**
-     * Add @c lhs transitions to @c lhs final states to the @c result.
-     */
-    void add_lhs_transitions_to_final_states();
-
-    /**
-     * Add @c lhs transitions from @c lhs final states.
-     */
-    void add_lhs_non_final_states_transitions();
-
-    /**
-     * Add @c lhs transitions to the @c result.
-     */
-    void add_lhs_transitions();
-};
+Nfa intersection(const Nfa &lhs, const Nfa &rhs);
 
 /**
  * Concatenate two NFAs.
