+++ src/nfa/nfa.cc
@@ -1070,24 +1070,6 @@ void Mata::Nfa::uni(Nfa *unionAutomaton, const Nfa &lhs, const Nfa &rhs) {
     }
 }
 
-void Mata::Nfa::intersection(Nfa *res, const Nfa &lhs, const Nfa &rhs, ProductMap*  prod_map) {
-    auto intersection { Intersection(lhs, rhs) };
-    if (prod_map != nullptr)
-    {
-        *prod_map = intersection.get_product_map();
-    }
-    *res = intersection.get_product();
-}
-
-Nfa Mata::Nfa::intersection(const Nfa& lhs, const Nfa& rhs, const Symbol epsilon, ProductMap*  prod_map) {
-    auto intersection { Intersection(lhs, rhs, epsilon) };
-    if (prod_map != nullptr)
-    {
-        *prod_map = intersection.get_product_map();
-    }
-    return intersection.get_product();
-}
-
 Simlib::Util::BinaryRelation Mata::Nfa::compute_relation(const Nfa& aut, const StringDict& params) {
     if (!haskey(params, "relation")) {
         throw std::runtime_error(std::to_string(__func__) +
@@ -1696,367 +1678,5 @@ void SegNfa::Segmentation::compute_epsilon_depths()
             visited[state_depth_pair.state] = true;
             process_state_depth_pair(state_depth_pair, worklist);
         }
-    }
-}
-
-void Intersection::compute()
-{
-    // TODO probably remove this since we use prod_map as parameter
-    //std::unordered_map<StatePair, State, decltype(hashStatePair)> thisAndOtherStateToIntersectState(10, hashStatePair); // TODO default buckets?
-
-    initialize_pairs_to_process();
-
-    if (lhs.trans_empty() || rhs.trans_empty()) { return; }
-
-    while (!pairs_to_process.empty())
-    {
-        pair_to_process = *pairs_to_process.begin();
-        pairs_to_process.erase(pair_to_process);
-
-        // TODO rewrite this (from previous Vata implementation -- rewrite the algorithm, or the format?)
-
-        compute_for_state_pair();
-    }
-}
-
-void Intersection::compute_for_state_pair()
-{
-    auto this_state_transitions_iter = lhs.transitionrelation[pair_to_process.first].begin();
-    auto other_state_transitions_iter = rhs.transitionrelation[pair_to_process.second].begin();
-    auto this_state_transitions_iter_end = lhs.transitionrelation[pair_to_process.first].end();
-    auto other_state_transitions_iter_end = rhs.transitionrelation[pair_to_process.second].end();
-    // find all transitions that have same symbol for first and the second state in the pair_to_process
-    while (this_state_transitions_iter != this_state_transitions_iter_end
-           && other_state_transitions_iter != other_state_transitions_iter_end)
-    {
-        // first iterator points to transition with smaller symbol, move it until it is either same or further than second iterator
-        if (this_state_transitions_iter->symbol < other_state_transitions_iter->symbol)
-        {
-            while (this_state_transitions_iter != this_state_transitions_iter_end
-                   && this_state_transitions_iter->symbol < other_state_transitions_iter->symbol)
-            {
-                ++this_state_transitions_iter;
-            }
-            if (this_state_transitions_iter == this_state_transitions_iter_end)
-            {
-                break;
-            }
-        }
-        else
-        {
-            // second iterator points to transition with smaller symbol, move it until it is either same or further than first iterator
-            while (other_state_transitions_iter != other_state_transitions_iter_end
-                   && this_state_transitions_iter->symbol > other_state_transitions_iter->symbol)
-            {
-                ++other_state_transitions_iter;
-            }
-            if (other_state_transitions_iter == other_state_transitions_iter_end) { break; }
-        }
-
-        // check both iterators point to the transitions with same symbol
-        if (this_state_transitions_iter->symbol == other_state_transitions_iter->symbol)
-        {
-            compute_for_same_symbols(*this_state_transitions_iter,
-                                     *other_state_transitions_iter);
-
-            ++this_state_transitions_iter;
-            ++other_state_transitions_iter;
-        }
-    }
-}
-
-void Intersection::compute_preserving_epsilon_transitions()
-{
-    initialize_pairs_to_process();
-
-    if (lhs.trans_empty() || rhs.trans_empty()) { return; }
-
-    while (!pairs_to_process.empty())
-    {
-        pair_to_process = *pairs_to_process.begin();
-        pairs_to_process.erase(pair_to_process);
-
-        // find all transitions that have same symbol for first and the second state in the pair_to_process
-        for (const auto& this_state_transitions: lhs.transitionrelation[pair_to_process.first] )
-        {
-            for (const auto& other_state_transitions: rhs.transitionrelation[pair_to_process.second])
-            {
-                compute_for_state_pair_preserving_eps_trans(this_state_transitions,
-                                                            other_state_transitions);
-            }
-        }
-    }
-}
-
-void Intersection::compute_for_state_pair_preserving_eps_trans(const TransSymbolStates& lhs_state_transitions,
-                                                               const TransSymbolStates& rhs_state_transitions)
-{
-    if (lhs_state_transitions.symbol == epsilon)
-    {
-        compute_for_lhs_state_epsilon_transition(lhs_state_transitions);
-    }
-
-    if (rhs_state_transitions.symbol == epsilon)
-    {
-        compute_for_rhs_epsilon_transition(rhs_state_transitions);
-    }
-
-    if (lhs_state_transitions.symbol == rhs_state_transitions.symbol)
-    {
-        compute_for_same_symbols(lhs_state_transitions, rhs_state_transitions);
-    }
-}
-
-void Intersection::compute_for_rhs_epsilon_transition(const TransSymbolStates& rhs_state_transitions)
-{
-    // create transition from the pair_to_process to all pairs between states to which first transition goes and states to which second one goes
-    TransSymbolStates intersection_transition(rhs_state_transitions.symbol);
-    for (State other_state_to: rhs_state_transitions.states_to)
-    {
-        create_product_state_and_trans(pair_to_process.first, other_state_to, intersection_transition);
-    }
-    add_product_transition(intersection_transition);
-}
-
-void Intersection::compute_for_lhs_state_epsilon_transition(const TransSymbolStates& lhs_state_transitions)
-{
-    // Create transition from the pair_to_process to all pairs between states to which first transition goes and states to which second one goes.
-    TransSymbolStates intersection_transition(lhs_state_transitions.symbol);
-    for (State this_state_to: lhs_state_transitions.states_to)
-    {
-        create_product_state_and_trans(this_state_to, pair_to_process.second, intersection_transition);
-    }
-    add_product_transition(intersection_transition);
-}
-
-void Intersection::create_product_state_and_trans(const State lhs_state_to, const State rhs_state_to,
-                                                  TransSymbolStates& intersect_transitions)
-{
-    StatePair intersect_state_pair_to(lhs_state_to, rhs_state_to);
-    State intersect_state_to;
-    if (product_map.find(intersect_state_pair_to) == product_map.end())
-    {
-        intersect_state_to = product.add_new_state();
-        product_map[intersect_state_pair_to] = intersect_state_to;
-        pairs_to_process.insert(intersect_state_pair_to);
-
-        if (lhs.has_final(lhs_state_to) && rhs.has_final(rhs_state_to))
-        {
-            product.make_final(intersect_state_to);
-        }
-    }
-    else
-    {
-        intersect_state_to = product_map[intersect_state_pair_to];
-    }
-    intersect_transitions.states_to.insert(intersect_state_to);
-}
-
-void Intersection::compute_for_same_symbols(const TransSymbolStates& lhs_state_transitions,
-                                            const TransSymbolStates& rhs_state_transitions)
-{
-    // Create transition from the pair_to_process to all pairs between states to which first transition goes and states
-    // to which second one goes.
-    TransSymbolStates intersection_transition(lhs_state_transitions.symbol);
-    for (State this_state_to: lhs_state_transitions.states_to)
-    {
-        for (State other_state_to: rhs_state_transitions.states_to)
-        {
-            create_product_state_and_trans(this_state_to, other_state_to, intersection_transition);
-        }
-    }
-
-    add_product_transition(intersection_transition);
-}
-
-void Mata::Nfa::Intersection::initialize_pairs_to_process()
-{
-    for (State this_initial_state : lhs.initialstates)
-    {
-        for (State other_initial_state : rhs.initialstates)
-        {
-            handle_initial_state_pairs(this_initial_state, other_initial_state);
-        }
-    }
-}
-
-void Mata::Nfa::Intersection::handle_initial_state_pairs(State lhs_initial_state, State rhs_initial_state)
-{
-    StatePair this_and_other_initial_state_pair(lhs_initial_state, rhs_initial_state);
-    State new_intersection_state = product.add_new_state();
-
-    product_map[this_and_other_initial_state_pair] = new_intersection_state;
-    pairs_to_process.insert(this_and_other_initial_state_pair);
-
-    product.make_initial(new_intersection_state);
-    if (lhs.has_final(lhs_initial_state) && rhs.has_final(rhs_initial_state))
-    {
-        product.make_final(new_intersection_state);
-    }
-}
-
-void Mata::Nfa::Intersection::add_product_transition(const TransSymbolStates& intersection_transition)
-{
-    auto& intersect_state_transitions{ product.transitionrelation[product_map[pair_to_process]] };
-    auto symbol_transitions{ intersect_state_transitions.find(intersection_transition) };
-    if (symbol_transitions == intersect_state_transitions.end())
-    {
-        intersect_state_transitions.push_back(intersection_transition);
-    }
-    else if (symbol_transitions->states_to != intersection_transition.states_to)
-    {
-        TransSymbolStates new_symbol_transitions{ intersection_transition.symbol,
-                                                  symbol_transitions->states_to.Union(intersection_transition.states_to) };
-        intersect_state_transitions.remove(*symbol_transitions);
-        intersect_state_transitions.push_back(new_symbol_transitions);
-    }
-}
-
-Nfa Mata::Nfa::concatenate(const Nfa& lhs, const Nfa& rhs)
-{
-     return Concatenate(lhs, rhs).get_result();
-}
-
-void Concatenate::compute_concatenation()
-{
-    map_states_to_result_states();
-
-    make_initial_states();
-
-    add_lhs_transitions();
-
-    make_final_states();
-
-    add_rhs_transitions();
-}
-
-void Concatenate::add_lhs_transitions()
-{
-    add_lhs_non_final_states_transitions();
-
-    add_lhs_transitions_to_final_states();
-
-    add_lhs_final_states_transitions();
-}
-
-void Concatenate::add_lhs_non_final_states_transitions()
-{
-    // Reindex all states in transitions in lhs, except for transitions concerning final states (both to and form final states).
-    for (State lhs_state{ 0 }; lhs_state < lhs_states_num; ++lhs_state)
-    {
-        if (!lhs.has_final(lhs_state))
-        {
-            for (const auto& symbol_transitions: lhs.get_transitions_from_state(lhs_state))
-            {
-                for (State lhs_state_to: symbol_transitions.states_to)
-                {
-                    if (!lhs.has_final(lhs_state_to))
-                    {
-                        result.add_trans(lhs_result_states_map[lhs_state],
-                                         symbol_transitions.symbol,
-                                         lhs_result_states_map[lhs_state_to]);
-                    }
-                }
-            }
-        }
-    }
-}
-
-void
-Concatenate::add_lhs_transitions_to_final_states()
-{
-    // For all transitions to lhs final states, point them to rhs initial states.
-    for (const auto& lhs_final_state: lhs.finalstates)
-    {
-        for (const auto& lhs_trans_to_final_state: lhs.get_transitions_to_state(lhs_final_state))
-        {
-            for (const auto& rhs_initial_state: rhs.initialstates)
-            {
-                if (lhs_trans_to_final_state.src == lhs_trans_to_final_state.tgt)
-                {
-                    // Handle self-loops on final states as lhs final states will not be present in the result automaton.
-                    result.add_trans(rhs_result_states_map[rhs_initial_state], lhs_trans_to_final_state.symb,
-                                     rhs_result_states_map[rhs_initial_state]);
-                }
-                else // All other transitions can be copied with updated initial state number.
-                {
-                    result.add_trans(lhs_result_states_map[lhs_trans_to_final_state.src], lhs_trans_to_final_state.symb,
-                                     rhs_result_states_map[rhs_initial_state]);
-                }
-            }
-        }
-    }
-}
-
-void Concatenate::add_lhs_final_states_transitions()
-{
-    // For all lhs final states, copy all their transitions, except for self-loops on final states.
-    for (const auto& lhs_final_state: lhs.finalstates)
-    {
-        for (const auto& transitions_from_lhs_final_state: lhs.get_transitions_from_state(lhs_final_state))
-        {
-            for (const auto& lhs_state_to: transitions_from_lhs_final_state.states_to) {
-                if (lhs_state_to != lhs_final_state) // Self-loops on final states already handled.
-                {
-                    for (const auto& rhs_initial_state: rhs.initialstates) {
-                        result.add_trans(rhs_result_states_map[rhs_initial_state],
-                                         transitions_from_lhs_final_state.symbol,
-                                         lhs_result_states_map[lhs_state_to]);
-                    }
-                }
-            }
-        }
-    }
-}
-
-void Concatenate::add_rhs_transitions()
-{
-    for (State rhs_state{ 0 }; rhs_state < rhs_states_num; ++rhs_state)
-    {
-        for (const auto& symbol_transitions: rhs.get_transitions_from_state(rhs_state))
-        {
-            for (const auto& rhs_state_to: symbol_transitions.states_to)
-            {
-                result.add_trans(rhs_result_states_map[rhs_state],
-                                 symbol_transitions.symbol,
-                                 rhs_result_states_map[rhs_state_to]);
-            }
-        }
-    }
-}
-
-void Concatenate::make_final_states()
-{
-    for (const auto& rhs_final_state: rhs.finalstates)
-    {
-        result.make_final(rhs_result_states_map[rhs_final_state]);
-    }
-}
-
-void Concatenate::make_initial_states()
-{
-    for (State lhs_initial_state: lhs.initialstates)
-    {
-        result.make_initial(lhs_result_states_map[lhs_initial_state]);
-    }
-}
-
-void Concatenate::map_states_to_result_states()
-{
-    State result_state_index{ 0 };
-
-    for (State lhs_state{ 0 }; lhs_state < lhs_states_num; ++lhs_state)
-    {
-        if (!lhs.has_final(lhs_state))
-        {
-            lhs_result_states_map.insert(std::make_pair(lhs_state, result_state_index));
-            ++result_state_index;
-        }
-    }
-
-    for (State rhs_state{ 0 }; rhs_state < rhs_states_num; ++rhs_state)
-    {
-        rhs_result_states_map.insert(std::make_pair(rhs_state, result_state_index));
-        ++result_state_index;
     }
 }
